dataflow/shareloader/requirements_marketstats.txt
apache-beam[gcp]
avro-python3
pyarrow
pandas-datareader
sendgrid
lxml
beautifulsoup4
vaderSentiment
finvizfinance
numpy
xlrd
openpyxl
google-genai
anio
httpx

